ref: Refactoring aplication
