+++ src/modules/accounts/infra/typeorm/repositories/UsersRepository.ts
@@ -1,8 +1,8 @@
+import { ICreateUserDTO } from "@modules/accounts/dtos/ICreateUserDTO";
+import { IUsersRepository } from "@modules/accounts/repositories/IUsersRepository";
 import { getRepository, Repository } from "typeorm";
 
-import { ICreateUserDTO } from "../../dtos/ICreateUserDTO";
-import { IUsersRepository } from "../IUsersRepository";
-import { User } from "../../entities/User";
+import { User } from "../entities/User";
 
 
 class UsersRepository implements IUsersRepository {
