+++ src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.ts
@@ -1,9 +1,9 @@
 import { inject, injectable } from "tsyringe";
-import { IUsersRepository } from "../../repositories/IUsersRepository";
 import { sign } from "jsonwebtoken";
-
 import { compare } from "bcryptjs";
-import { AppError } from "../../../../errors/AppError";
+
+import { AppError } from "@shared/errors/AppError";
+import { IUsersRepository } from "@modules/accounts/repositories/IUsersRepository";
 
 interface IRequest {
     email: string;
