+++ src/modules/cars/infra/repositories/SpecificationsRepository.ts
@@ -1,9 +1,9 @@
 import { getRepository, Repository } from "typeorm";
-import { Specification } from "../../entities/Specification";
+import { Specification } from "../typeorm/entities/Specification";
 import { 
     ISpecificationsRepository, 
     ICreateSpecificationDTO,
-} from "../ISpecificationsRepository";
+} from "../../repositories/ISpecificationsRepository";
 
 
 class SpecificationsRepository implements ISpecificationsRepository {
