+++ src/modules/cars/repositories/ISpecificationsRepository.ts
@@ -1,4 +1,4 @@
-import { Specification } from "../entities/Specification";
+import { Specification } from "../infra/typeorm/entities/Specification";
 
 interface ICreateSpecificationDTO {
     name: string;
