+++ src/modules/cars/useCases/importCategory/ImportCategoryUseCase.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from "tsyringe";
 import fs from "fs";
 import { parse as csvParse } from "csv-parse";
-import { ICategoriesRepository } from "../../repositories/ICategoriesRepository";
+import { ICategoriesRepository } from "@modules/cars/repositories/ICategoriesRepository";
 
 interface IImportCategory {
     name: string;
