+++ src/shared/container/index.ts
@@ -1,13 +1,13 @@
 import { container } from 'tsyringe';
 
-import { ICategoriesRepository } from '../../modules/cars/repositories/ICategoriesRepository';
-import { CategoriesRepository } from '../../modules/cars/repositories/implementations/CategoriesRepository';
+import { ICategoriesRepository } from '@modules/cars/repositories/ICategoriesRepository';
+import { CategoriesRepository } from '@modules/cars/infra/repositories/CategoriesRepository';
 
-import { ISpecificationsRepository } from '../../modules/cars/repositories/ISpecificationsRepository';
-import { SpecificationsRepository } from '../../modules/cars/repositories/implementations/SpecificationsRepository';
+import { ISpecificationsRepository } from '@modules/cars/repositories/ISpecificationsRepository';
+import { SpecificationsRepository } from '@modules/cars/infra/repositories/SpecificationsRepository';
 
-import { IUsersRepository } from '../../modules/accounts/repositories/IUsersRepository';
-import { UsersRepository } from '../../modules/accounts/repositories/implementations/UsersRepository';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { UsersRepository } from '@modules/accounts/infra/typeorm/repositories/UsersRepository';
 
 container.registerSingleton<ICategoriesRepository>(
     "CategoriesRepository",
