+++ src/shared/infra/http/server.ts
@@ -3,14 +3,14 @@ import express, { NextFunction, Request, response, Response } from "express";
 import "express-async-errors";
 import swaggerUi from "swagger-ui-express";
 
-import "./database";
+import "@shared/infra/typeorm";
 
-import "./shared/container";
+import "@shared/container";
 
 import { router } from './routes';
-import swaggerFile from './swagger.json';
+import swaggerFile from '../../../swagger.json';
 
-import { AppError } from './errors/AppError';
+import { AppError } from '@shared/errors/AppError';
 
 const app = express();
